+++ js/data.js
@@ -1,3 +1,4 @@
+import { getRandomElementArray, getRandomNumber } from './util.js';
 const COUNT_OBJECTS = 25;
 const DESCRIPTIONS = [
   'Музыка',
@@ -24,4 +25,21 @@ const CountComments = {
   MAX: 200
 };
 
-export { COUNT_OBJECTS, DESCRIPTIONS, CountLikes, CountComments };
+const createPhoto = (id) => ({
+  id,
+  url: `photos/${id}.jpg`,
+  description: Array.from({ length: getRandomNumber(2, 3) }, () =>
+    getRandomElementArray(DESCRIPTIONS)
+  )
+    .map((x) => `#${x}`)
+    .join(' '),
+  likes: getRandomNumber(CountLikes.MIN, CountLikes.MAX),
+  comments: getRandomNumber(CountComments.MIN, CountComments.MAX)
+});
+
+const createArrayObjects = () =>
+  Array.from({ length: COUNT_OBJECTS }, (_, photoIndex) =>
+    createPhoto(photoIndex + 1)
+  );
+
+export { createArrayObjects };
