+++ js/util.js
@@ -1,10 +1,3 @@
-import {
-  DESCRIPTIONS,
-  CountLikes,
-  CountComments,
-  COUNT_OBJECTS
-} from './data.js';
-
 const getRandomNumber = (firstNumder, secondNumder) => {
   if (firstNumder < 0 || secondNumder < 0) {
     return NaN;
@@ -16,23 +9,7 @@ const getRandomNumber = (firstNumder, secondNumder) => {
 
 const isMaxLengthStr = (str, maxLength) => maxLength >= str.length;
 
-const getrandonElementArray = (array) =>
+const getRandomElementArray = (array) =>
   array[getRandomNumber(0, array.length - 1)];
 
-const createPhoto = (id) => ({
-  id,
-  url: `photos/${id}.jpg`,
-  description: Array.from({ length: getRandomNumber(2, 3) }, () =>
-    getrandonElementArray(DESCRIPTIONS)
-  )
-    .map((x) => `#${x}`)
-    .join(' '),
-  likes: getRandomNumber(CountLikes.MIN, CountLikes.MAX),
-  comments: getRandomNumber(CountComments.MIN, CountComments.MAX)
-});
-
-const createArrayObjects = () =>
-  Array.from({ length: COUNT_OBJECTS }, (_, photoIndex) =>
-    createPhoto(photoIndex + 1)
-  );
-export { createArrayObjects, isMaxLengthStr };
+export { getRandomElementArray, isMaxLengthStr, getRandomNumber };
